Move freetype collection & link in separate files
